🧹 Extract monolithic cache.py into smaller modules
