+++ pr_description.md
@@ -1,0 +1,10 @@
+🎯 **What:** Extracted the monolithic `cache.py` (over 1000 lines) into a modular `cache/` directory containing logically separated files (`base.py`, `local.py`, `distributed.py`, `hybrid.py`, `feature.py`, `utils.py`). The existing structure and backward compatibility are maintained through `cache/__init__.py`.
+
+💡 **Why:** Files over 500 lines typically contain multiple responsibilities. Breaking down the cache module into its constituent parts (base, specific backends, features, and utils) makes it significantly easier to navigate, maintain, and expand the caching capabilities of `settler_workhorse` without navigating a monolithic file.
+
+✅ **Verification:**
+- Validated correct behavior by successfully running the full `pytest` suite for `packages/workhorse`.
+- Verified identical function structure, variables, and comments through direct exact string extraction (using `sed -n 'X,Yp'`) from the original `cache.py`.
+- Formatted and linted the refactored files using `ruff`.
+
+✨ **Result:** A more robust, readable, and maintainable caching layer in `settler_workhorse` mapped directly to specific cache implementations.
